Add files via upload
Intelligent Dustbin: An automated dust collecting and notify system 
Duration: November 2021
Description: It’s an intelligent dustbin which is open when a man arrived and wants to 
through a dust inside it. When the dustbin is full a message is sent to the owner number via 
SMS system. It also indicates the dust level that are 25%, 50% and 100%.